23.06 Added changes to the echo-bot
