+++ development/bot/bot.py
@@ -1,0 +1,5 @@
+class Bot:
+    
+    def echo(self, message):
+        self.message = message
+        print(f'Bot: {self.message}')
